Framework update 3 - fix the damn thing in ExoticaTechHM
Well, 3am coding syndrome struck again, and somehow I thought I could manage
the edge-case of "affecting all modules but not sharing with them" in one
simple if() statement. By extending the old if() statement. Which turned out
rather inconvenient, or rather - downright impossible, and I didn't want
to have two "if(...) continue" one after another since they were still largely related.

So, the if() logic was extracted to a new shouldSkipModification() method(s) with the extra
logic added there. Now, GuardianShields no longer multiply as they did before.

This is it, the issue is *fixed* unless it was a problem with a bad build, like sometimes
Conditions fail to work on 'bad builds' for some weird reason.

In any case, the only thing left to do now regarding this issue, is to prevent further
installations with a new Condition and add it to the systems that should allow only one
instance on the ship - as adding multiple instances will essentially yield zero effect or
end up breaking something.