Fix the NotIn condition operator
It was using the same "in" key as the In condition operator.
Fortunatelly, neither of these is actually being used, the 'In' operator is used in just one place
that also doesn't seem to work so... that's something to be fixed later on.

For what it's worth, the SpooledFeeders is requiring that "Safety Overrides" hullmod be installed,
yet that doesn't seem to work at all.