Console extension part1 - A new et_additem command
So, a bit more context on this.
I've been made aware through discord that exotic items (which are an additional
InstallationMethod for exotics) are actually superpositioning their exotica icon
inside the exotic item's icon. I really dislike this.

So, to help test this, and give players more cheating capabilities, I've made a new
console command to grant these exotic items. But alas, due to 'java.io.*' package(s)
being prohibited, I can't make this decently enough to actually read the special_items.csv
file, parse it, and then extract the item IDs from there.

So, the next best thing will have to do - whenever a new item is added, the help for
et_additem will have to be ammended in ET/data/console/commands.csv with the ID of the
new item and which system it's for.