RewindSystem polishing part1 - add text next to it explaining why it can't activate
Title says it all. At first I started changing alot of stuff to try and fix a dangling issue
which was occuring in the "previousStates.find { shipParams -> " line when I was getting
crashes due to non-null shipParams being null, and then I introduced Optionals into the
RingBuffer but none of that was really necessary, all that was really necessary was to
just declare the shipParams as nullable.

So I reverted all of the Optional changes in the RingBuffer and it's tests because that
still didn't quite get rid of the "non-null shipParams being null" issue, and stuck
with a simpler solution. If this one doesn't quite fix it either well then, Optionals
can always be brought back into it later on.

At the time being, RewindSystem is printing out additional info about why it can't activate,
and there were a few other small pedantic thinks that were taken care of as well. It works well,
and seems stable.