RewindSystem polishing part2 - significantly improved the Afterimage effect
Apart from applying the Afterimage to the whole ship rather than just the module now,
now there are two Afterimages, one above the ship which combines with the sprite color
and jitters a little bit, and one below the ship which doesn't combine with the sprite
color and jitters twice as much. The end result is very pleasing to the eye.

Can't believe it, but the RewindSystem really came together quite well.