RewindSystem polishing part3 - added gliding Afterimages
Just before the ship teleports, it's going to spawn two more afterimages that will
float away towards the destination point. Since they can't cover that whole area
because there's obviously some kind of limitation, perhaps another afterimage
right after the ship teleports can be made, from "somewhere" but so it lands on
the position where the ship is now might make for a better look.

So far, it's looking alright, and somewhat better than it used to. Just one small
detail left to completely round it up.

EDIT1: alright, the "one small detail" is implemented, and the POC looks somewhat alright
now, there's more work to be done on this class to make it nicer and to revert the old
'after teleport' afterimage parameters back

EDIT2: necessary refactoring done, extracted all Afterimage creations to separate methods,
however one problem still remains - after the ship teleports, the "after teleport" after images
come in with a delay instead of immediatelly.

TODO: get rid of systemActiveTimeInFrames and clean up the logtag list; advance() doesn't return
frames but rather seconds. and the system gets activated and 4 seconds later it turns off.

EDIT3: got rid of the delay, for some reason it didn't work in the else() branch of the
if (systemActivated.get), or one of them at least...
The solution was to introduce a (yet another) flag which gets set as soon as the ship is teleported
via deploySavedState() and this new 'spawnAfterTeleportAfterimages' flag is read directly in the
advance() and unset accordingly.

Either way, now the problem is - the pre-teleport gliding afterimage also comes along with us
and is visible. So speed it up so that even when you teleport it's out of sight.

EDIT4: the pre-teleport afterimage problem fixed, updated string to mention it's only installable
on main module, added tests for the new TimeUtils, also added latest JAR with these changes.