Fix a bug in ExoticaTechHM - fix NPE on fleet member
The code seemed to only cover a "happy path", when the ship.getFleetMember() was non-null,
and subsequently checking ship.getFleetMember().getShipName() to be null.

However, in the case when ship.getFleetMember() *was* null, trying to check the ship's name
hits into a NullPointerException.

Fixed it so that if the fleetmember is null, the 'fleetMemberShipNameIsNull' defaults to false.