RewindSystem polishing part 4 - making it work on weapons and other modules
Since it currently only operates on the module it's installed on (main module),
it falls short of it's original vision of restoring the *whole* ship's
health / velocity / direction / position / cooldowns as well as all weapon data.

However, this first commit ensures that all of what is needed actually works on
the main module. So now it just needs to be generalized enough throughout the whole
ship (all of it's modules) and it's good to go.

EDIT1: Initial multi-module implementation and a few touchups to single-module one.
The fact that i have two implementations bothers me, but now the shields and systems
are also being checked for.
I need to streamline this asap, since i already had one whoopsie slip through due to
copypasting at 4am...
Consider this a 'checkpoint' step.

EDIT2: refactored and simplified code quite a bit.
- isActivationPossible() is greatly simplified by doing the same check for ConcreteShipParams
and MultiModuleShipParams while returning 'false' for EmptyShipParams and nulls
- deploySavedState() is greatly simplified by delegating most of the savedState deploying to
a new method - applySavedStateToModule(module, savedState)

EDIT3: RewindSystem so far works alright and has feature-parity for single-module and multimodule
ships. Also fixed a terrible bug in Extensions regarding the getAllShipSections() which never
included the main-module for multi-module ships, and added a few other utility methods as well.
The latest JAR is in here too now.

EDIT4: Edit the string to mention that it does not revive destroyed modules, added latest JAR.