Did a pass on upgrades too
All four upgrades now have their strings streamlined like the rest of them, and mention
whether they work on the whole ship or just the module they're installed on. They also
mention whether they're installable on just the main module or not.

Finally, Kingslayer was changed to share it's effect to the whole ship, since it seems
it's very rare.