Refactor and simplify the addAfterimageToWholeShip() method
Now I know why i never caught that the getAllShipSections was broken!
It was broken off this method, but this method cleanly applied it to all
sections (like addAfterimageToWholeShip() returns) and on top of that it
applies the afterimage to the passed ship as well;

either way, now the addAfterimageToWholeShip() uses getAllShipSections() internally.

EDIT: improved javadocs