SpooledFeeder conditions revival and add support for disjunct conditions
Since all conditions are conjuct (AND-ed) and they all have to be met for
a system to be installable, add support for disjunct (OR) conditions by
introducing a 'minConditionsNeeded' key in exotics.json and add support
for it in Modification. This makes it so that if we have 2 conditions in the
conditions list, and a minConditionsNeeded of 1, it effectivelly ends up being
condition1 OR condition2 when it really is "one of these two".

Initially, it *looks* like it checks out, but finer-grained testing is required
since:
1) none of the 'cannotApplyReason's are being printed, instead of saying <reason> OR <reason>
2) there seem to be cases when a ship that shouldn't meet the conditions seem to be meeting them

EDIT: Both problems fixed.
were doing their job properly, they were futile/useless since in the end we just used count() and counted
*everything* rather than just elements that were true, or filtering the ones out, and so on.
So the last ".map { weaponSpecAPI -> weaponSpecAPI.type == BALLISTIC }" was just replaced with "count { }"
Same predicate and everything, but this works fine.

However, there's a new problem now - the multiple 'cannotApplyReason' text is going over the buttons.

This is a 'checkpoint' commit again, before i end up breaking things; at least i'll have something
to come back to.