Fix the new UI problem caused by disjunct conditions (SpooledFeeders)
Made somewhat reactive UI, it's not the greatest or really reactive but
it does react to the number of 'cant apply label' list size.

Try to center the OR somehow.

EDIT: put an unused 'lastLabel' to use but i could have just deleted it...

EDIT1: ok, now it's kinda centered, but it's centered on the tooltip's width
not the last label's width. consider changing it up.