Revert NanotechArmor, PenanceEngine and PlasmaFluxCatalyst
Since they do their magic in applyToStats() which is called from
beforeShipCreation() (so they can be visible/usable in campaign)
it makes no sense to fight this and I won't be able to find a decent
solution to making their host-module share these effects to other modules,
since - besides in Refit screen - modules aren't part of the ship, or rather
FleetMember.

So it's just easier to get rid of the misleading strings and get rid of the
shouldShareEffectsToOtherModules() implementations from them since it's not doing
anything good.

Yes, I'm taking the easier way out.

EDIT: removed the implementations, added fresh JAR