Checkpoint commit - installing Alpha Subcore now properly affects all modules' stats
So, half way there. The following problems remain:
1) when nothing is installed on other modules, and alpha subcore (and it's hullmod) are installed,
there's no feedback about that at all.

So the 'exoticatech' hullmod should probably be added to display it.

2) uninstalling (destroying) the alpha subcore probably leaves stats in their improved state.
That should be looked into.

3) Exotic-related Hullmods (not HullmodExotics) shouldn't directly inherit BaseHullmod.
Another layer of indirection would be great there, so that we could have a hierarchy such as:

AlphaSubcore (hullmod exotic) references AlphaSubcoreHM (hullmod)'s ID, which needs to match the ID
of the entry from hull_mods.csv file. Better than just leaving string literals sprinkled everywhere.

So, so far seems like an improvement, but there's still work to be done on this.