Enable the 'installable only on main module but works on whole ship' usecase
And also turn off AlphaSubcore affecting modules too, since installing it on
non-main modules will likely not work for any other module either (untested yet).

Either way, now it's back to it's original idea - installable only on the main module
and working on the whole ship.

I just need to tackle the problem of the OP costs not returning to what they were when
it's uninstalled and bam, will be great then. Until then, it just sucks when it's uninstalled.