Add synchronization to GuardianShield and StringUtils
I'm really not sure what this will accomplish (without the underlying thing being `volatile`)
but for the time being - let it rest like this. It can't hurt.

But might be completely unnecessary.

The problem to be solved is - hullmod exoticas MutableStats not returning back to what they were
before AlphaSubcore was installed.