Cleaning up HullmodExotic and HullmodExoticHandler
As initially planned, the "internal" methods that noone other than
HullmodExoticHandler should be touching have been made private (and somewhat obsolete
but that's a different thing)

The HullmodExotic was just changed to call into the HullmodExoticHandler.removeHullmodExoticFromFleetMember
if we obtained an instance of the ExoticHullmod from the ExoticHullmodLookup class.

So far, working good still.
EDIT: updated some inline comments to explain what's going on better