Work on reducing (fixing) the infinitely-growing map problem
So, the current state of things are - when teh alpha subcore is installed, the lookup map contains only one entry.
Which is great.
But somehow, after removing the alpha subcore, it's size jumps to 5 ......

but wait a minute.
I never looked at the size of the keySet or the number of entries in the debugger.

Still, step through the removeHullmod(...) method with a debugger and see *when* and *why* it grows so much,
or if it's just a false-positive or not.