Fix the logical error of 5am code ...
Ok, so the thing that kinda "broke" both the refit and the default 'exotica' colony screen
was that I was expecting the variant to be expected before uninstalling, but since at some point
I also decided to **stop** expecting it once it's installed on that variant, this broke it, so things
never uninstalled.

Now, with the default 'exotica' colony screen fixed and working GREAT, i also need to fix the refit screen
(or both?) which has the problem of not being able to install again due to already having HullmodExoticInstallData
but the variants not being expected.

So obviously, in the end of onDestroy() I also need to get rid of the InstallData key to allow for reinstalling.
Would be great if that fixed the refit screen too :)