Fix tests and the HullmodExoticHandler garbage carrying over between games
As was previously mentioned, the refit screen does throw in alot of garbage fleetmembers
in there which just leave their keys in the map. The whole point why this PR was started.

However, since that's a non-issue, AND REALLY NOT BELIEVEING ALEX'S UID to actually be unique
(probably unique on a single game level but totally not unique) lets also clear the map when
a map is started/loaded and have the map regenerate normally from there.