Started extracting the "flows" from HullmodExotic to HullmodExoticHandler - part 1
So far, the two "shouldInstall" / "install" flows have been moved/extracted from the HullmodExotic
to the HullmodExoticHandler, with callbacks for each step.

So far, the installation still works fine for "exotica tech" colony screen and UNTESTED for refit screen,
but is most likely still broken on the refit screen.

This is something that will be handled internally by the HullmodExoticHandler and calling different things
by using it's new "runningFromRefitScreen()" method.

Most likely, different implementations of "should" / "do" methods for both installing and removing will have to
be provided, the strict version which is what we have now, and a new more lax version for Refit scren which throws
in alot of potential problems with it's faked FleetMemberAPIs for modules.

Still, so far it seems to be heading in a right direction, and keeping the HullmodExotic much 'lighter' and simpler
with HullmodExoticHandler doing all the heavy lifting.

Once we get to feature-parity between the exotica colony screen and refit screens, we can start simplifying the HullmodExoticHandler
as well. Because, e.g. the part about getting child modules (or all *other* parts of the ship/FleetMemberAPI variant) can be done
with an already-existing extension method; and I have no good ideas about getting other parts of teh ship from non-root module
part (or variant) like it's pretty straightforward to do from a ShipAPI.

Worst case scenario - we'll just warn people not to install AlphaSubcores from Refit screen. But since I also want to do that,
and it used to work - makes me push for this still.