Extracting of "flows" from HullmodExotic to HullmodExoticHandler - part 2
The "shouldRemove" / "remove" flows have been moved/extracted from the HullmodExotic
to the HullmodExoticHandler as well.

As expected, it works fine for the "exotica tech" colony screen and is broken for refit screen,
installing only on main module (disregarding child modules) and doesn't quite uninstall it
from the main module either.

Left TODOs on what to change for the LaxWorkPerformer to be used by Refit Screen and the StrictWorkPerformer
to be used by the "exotica tech" colony scren.

The key difference being (what I suspect/expect) is the removal from 'expected' variants, since the fake
fleetmembers happening from the refit screen don't quite pass the 'expected' filter more than once (as they did while it was working)
so this needs to be taken care of *somehow*.

In fact, taking care of the LaxWorkPerformer vs StrictWorkPerformer implementations is going to be easy.
Ensuring their compatibility however, is going to be the harder part of this.

Because we also want to satisfy the following usecase(s):
1) Install from refit
2) it working fine across all modules (all modules having spare OP)
3) uninstalling from exotica screen
4) it working fine across all modules (no modules having spare OP)

1) install from exotica screen
2) it working fine across all modules (all modules having spare OP)
3) uninstalling from refit screen
4) it working fine across all modules (no modules having spare OP)

When we get there at that feature-parity moment, we can start optimizing away garbage from the individual flows / WorkPerformer implementations.