REVERTME - checkpoint commit before i try fixing the problem
I think I see the actual problem... like, the real root cause of it.
The "check and remove from all child modules" doesn't even look at the
"installed modules" list we saved in the lookup map, it just fetches the child
modules again from scratch.

And *that's* most likely the culprit.

So, since this commit is just a checkpoint, and does nothing good/useful besides
shuffling a few commands here and there... lets try fixing this problem first
with a decent checkpoint to get back to, even if this one ends up being reverted.