Fix uninstalling from Refit screen - the "CheckAndRemoveFromAllChildModulesVariants" flow was broken
It was literally because I "started" the function by copypasting one of the 'install' flows and
modified it from there - while completely not seeing that the module variants were being directly fetched
from the FleetMemberAPI Variant we were passing.

Which is obviously no good, because in refit screen we end up installing on what, 10-15 modules for a 3-module ship
and once we uninstall from only 2 - that's where the "leftover" was.

Installing/uninstalling from Exotica screen works fine, Installing/uninstalling from Refit screen works fine.

I HAVE NOT TRIED MIXING THE TWO - but... maybe it works, maybe not.
If it works - great. ggwp.
If it does not work - maybe the InstallData itself will have to be slightly modified.

But so far so good, can proceed with the cleanup and codereview and eliminating unnecessary crap.