Get rid of unnecessary stuff and re-test the cross scenarios
Well, the cross-test didn't quite go super well.
Installing in Refit and Uninstalling in Exotica Screen works fine.
Installing in Exotica Screen and uninstalling in Refit doesn't quite work fine.

That case will probably need a warning, or I just let people screw up their ships
and re-fix them by installing/uninstalling again from the same place.

Either way, I'm happy that both places work, and even one half of the cross-test works fine.

I don't think i'll be pursuing this to the end because, as it is (and as I've come to understand it),
the Exotica->Refit combo isn't going to fly because we'll be missing all those fake FMAPIs that get
spawned in by the refit screen.