Fix addAfterimageToWholeShip and add considerate visual flair to ExternalThrusters
So, the big problem with the addAfterimageToWholeShip from Extensions was that, even after it obtained
all ship's sections, it was still applying the afterimage to *just the ship* instead of the section.

This was surely causing the afterimage to be applied on the same main module section for every section,
instead of just applying it to the section we're currently looping over.

Besides that, after discovering this issue, i got an idea to try out the ExternalThrusters since
I got a great idea of "what if i could make the flames alot bigger when it starts up" (since that's how I
discovered the issue of not applying the afterimage to the whole ship to begin with) and... I added it.

And it looks damn great. Camera shake would indeed make it better, but - until then... this is good enough.