REVERTME - Make AlphaSubcore installable only on main module and update it's string
To avoid the "accidental" crashes that will be caused by people installing this properly and then
trying to uninstall it from submodules, lets temporarily make this installable only on the main module.

Thing is, so far it seems to work great.

However, I would also like this to **NOT** remain the final solution. Since the crashes come due to throwing
so lets see what happens when that one gets removed as well.