CR comment - get rid of elvis operator and throwing when InstallData is not found
Since trying to remove the AlphaSubcore (installed on main module) from child modules would always throw due to not finding
the InstallData for the faked up FleetMemberAPI for the child member, I tried restricting them to just the main module.

However, when the throwing was replaced with just logging an error and returning, even though it "seems" as if the AlphaSubcore
was removed from the child (and the ExoticaTechnologies hullmod was removed !!!) after flipping between modules everything is still "fine"
and it was not really removed or the additional AlphaSubcore added to the inventory.

This is.... somewhat fine?