Add the TODO/FIXME to HullmodExotic to mention fixing the "effect sharing"
Basically, making it so that any module can share to whole ship.
For the time being, only "root" modules can share to all child modules.

But the "lookup" for even finding the child modules to grab it's root module
and proceed to grab all other modules from there is... not super complicated,
but with the initial idea I have to lookup through the fleet and all - it's suboptimal.

The check (and flow) should be different depending on whether the module is a child module
or a root module. And the fact that child modules' FleetMemberAPIs are faked at-request,
it's gonna make the ExoticHullmodLookup entirely different; since they'd first need to become
stable for this to work.

I don't want to let this go, however I see no good reason for this not-currently-used or needed usecase
to block further development or release, given that there's what, only 2 hullmod exotics and I have no
real plans to keep on making *them* vs other Exotic systems.