Split DaemonCore from AlphaSubcore - part1
While most of the split is "done", there is still something seriously busted here.

Namely, when a HullmodExotic **doesn't** share effects to all children, the changes are neither visible
on the refit screen, nor do they work properly from the refit screen; however they *do* seem to work alright
from the exotica colony menu screen, so something isn't quite right in either the HullmodExotic or the HullmodExoticHandler

Which needs further polish.

But for now, this part of the split is done. The rest needs testing and tinkering with.