HullmodExotic fixes - wip
Turns out, when the HullmodExotic doesn't share it's effects - nothing worked!
Somehow, the shareToAllModules aka the "install to all child modules" made everything work.

Part of this problem was on the HullmodExoticHandler's side, but a part of it is also on the
HullmodExotic side of things as well.

So, saving this, before I start minifying the changeset.
For now - installing/uninstalling works fine from refit for hullmod exotics that don't share their effects.