HullmodExotic fixed and DaemonCore is good to go
First off, I'll answer my previous conundrum regarding "suspicious" need to have the
uninstallation performed from the refit variant as well - it is **not** necessary when
the HullmodExotic is sharing effects to other modules.

HOWEVER it is **very** much necessary when it is not sharing effects.
Thus, the previous "suspicious fix" is really needed.

Once the DaemonCore was switched to share effects to all other modules, it wasn't needed anymore.
But i'm keeping it because it's causing zero problems, and is there for non-sharing hullmod exoticas.

Apart from that, finally commit the extension method used in HullmodExotic to check if we're in Refit screen,
and fix the DaemonCore's string to mention it's working on the whole ship and is only installable on the main module.