Fix the GuardianShield crash
It was caused by a rather mystic error - FormatFlagsConversionMismatchException: Conversion = s, Flags =

which... hints that there was a "% s" somewhere, at best.

And after searching a bit, accepting that I was hopefully not that stupid but maybe were somewhere -
I found the culprit.

Caused by GuardianShield's longDescription string.
 =Stabilized Shields=: +100% shield push

This little bit stopped crashing once it was changed to
 =Stabilized Shields=: +100%% shield push

So, the funny crash that was pestering me since, well, maybe a very long time ago - is finally fixed.