Fix HullmodExoticHandler so it properly installs HullmodExotics on single-module ships
Well well well, since i was so focused and obsessed with multimodule ships, i didn't really try it out on singlemodule
ships until it was "too late" which is when a nasty NPE hit me in the face - turns out, variant.getStatsForOpCosts was very
fine to use for multimodule ship's modules. However, on singlemoduled ships, it seems to always be null?

So a new method was made, that returns the FleetMemberAPI's stats if variant's statsForOpCosts are null.

And that seems to work perfectly well, the AlphaSubcore can be installed/uninstalled, the OP prices go down then back to what they were.
Everything works out great.

Hooray me.