Small touchup to ExternalThrusters and latest JAR
While I wanted to make the ExternalThrusters do a small "screen shake" and parametrize it via luna settings,
the inability (skill issue?) to properly implement it so it works as if the viewport is centered on the ship
while applying a small jitter doesn't quite work so well (and robs the user of being able to move the viewport around
or zoom in/out) it's not a super great feature.

However, the small touchup is that ExternalThrusters now apply a constant acceleration forward while the system is active.

Also added latest JAR