Another touchup to ExternalThrusters, and latest JAR
Well, after adding the bit to force acceleration, i wondered if it's gonna end up
going in maybe a whole complete wrong direction, so i wanted to add a check to make
sure it doesn't activate if it's not within a certain arc in front of the ship.

Luckily for me - that part was already there, i thought about it on time.
But there is one small problem with previous implementation:

it would check if the distance is over 2000 and if it's within a absolute 15 degree
(technically 30 degree) arc ahead of the ship and activate then, otherwise it'd activate
only if it was far enough.

That "otherwise" is no longer good, because the ship would be unable to turn while activated
anymore.

So now it'll activate only if it's more than 2000 range away and within a 30-degree arc from the
ship's forward vector, and not activate otherwise.

Also, latest JAR with the change.