Patch HullmodExotic::unapplyExoticHullmodAndRemoveExoticaAndHullmod()
Turns out, some multi-module ships *also* run into the exact same problem that the singlemodule ship(s) did.
No wonder - the ork dreadnought one was to blame.

Regardless, the exact same cure works; falling back to fleetMember.getStats() if variant.getStatsForOpCosts()
is null.

And to avoid copypasting that method, or copying it to util/extensions, I made it 'internal' in HullmodExoticHandler
because why not just have it there and be accessible for hullmods and hullmod exotics.

This small change makes the ork dorn dhe ship work fine and installing/uninstalling AlphaSubcore on it also works fine
and reverts OP costs across all modules as-expected.

Nice.