Reproed the 'flying shield with no enemies left' problem ...
So, obviously, the previous solution isn't good at all. It will either have to be refactored or completely
done from scratch. For now, lets try plugging it like this and watching if it reproes again.

Still, this should be refactored at least, with the whole "drone removing" to be extracted into a separate
"onHostShipRemovedOrDestroyed()" method and then just calling *that* from all these other places ...