Consolidate max number of exotics for multimodule ships, at least for the player's side ...
So, it turns out that submodules (child modules) were being limited to 2 exotics, while the main module
could have 3 (due to commander having the "best of the best" skill learned).

Root cause of this being - submodules's fleetData is null and *isn't* cloned from the parent's FleetMemberAPI.

So, in `ShipModifications`, for player-owned ships, we won't be looking at just `member.fleetData`;
in case when that is null, we will look at player's fleet (obtained via `Global.getSector().playerFleet`)
and then let the FactionConfigLoader get the FactionConfig for the playerFaction's ID, and get the max exotics number from there.

In the weird odd case when the player's fleet is null, we'll fall back to the `getMaxExoticsNumber()` which simply returns `ETModSettings.MAX_EXOTICS`

The `FactionConfigLoader` will no longer directly default to `ETModSettings.MAX_EXOTICS` but rather will delegate that to a method,
so that the MAX_EXOTICS value is reevaluated on every call of the method.

The `FactionConfig` got the biggest overhaul - while it's effectivelly the same as it was, it wasn't working at all for multimodule ships.
While it's going to be plaky for enemy multimodule ships (especially their child modules') at least it's now properly fixed for the player's side.
Problem, of course, being that child modules' neither have fleetData, nor fleetCommander which we could query for having the "best_of_the_best" skill.

So, for player owned ships, we will simply look if the commander of the player's fleet (obtained via `Global.getSector().playerFleet`) has the skill.

This means that enemy's mutlimodule ships' child modules will have a "max exotica number" limit less than the main module in case it's fleetCommander
does have the skill.

But I can live with that.

Now onto to addresss the other problem - maxing out exoticas on submodules, then installing alpha subcore on main module, then preventing it from
somehow ~~clearing~~ replacing exotics on child modules, and clearing everything when uninstalled.