Fix HullmodExotic problem discovered while testing out the "over max exoticas limit" edge-case
The problem in question was, once a max number of exotics was installed, installing the AlphaSubcore
would somehow "overwrite" all installed exotics on child modules. Uninstalling the AlphaSubcore would
remove *everything* that was installed.

Obviously, this should not be the case - if the child modules had upgrades/exotics installed, installing
the AlphaSubcore should simply **add** it to the list, and then uninstalling it should just remove it from
the list, not removing the "exoticatech" ET hullmod if there are leftover modifications remaining on the module.

This has now been fixed.

But one more problem remains - when max number of exotics is installed on child modules, installing the AlphaSubcore
on the main module still installs it on all children, exceeding the limit. This needs to be fixed to avoid "cheating"
the system for one exotica and avoid going over the limit.