Revert changes to HullmodExotic
Well, since I broke it, this needs to be more carefully revised.
Or just left as-is.

If the "alpha subcore uninstalling cleans up all exoticas" issue is there, it needs to be tackled.
But for now - better to have AlphaSubcore installing/uninstalling work from both exotica and refit,
than not working.

Worst case scenario - I'll just revert this commit.