Fix the ExoticaTech hullmod problem that creeped in
Alright so as stated before - installing AlphaSubcore on an empty ship
would make the ExoticaTech hullmod show up on main module and **not** show up
on child modules.

This was due to flawed logic, which assumed that when we get to the
installThisHullmodExoticToFleetMembersVariant() method, it was *expected* that
no installations happened yet, and that ExoticData was actually empty.

It wasn't, ExoticData is very much not empty at that moment. The method got called
*after* it was installed, so the whole new idea of "lets apply the hullmod only if we
don't have mods installed" is flawed by design.

So, it was updated to not act smart about it, since "installing" the ExoticaTech hullmod
will either remove it or add it or update it.

Test results:

BEFORE:
installing AlphaSubcore on a ship without modifications would make the hullmod show only on
main module and not show anywhere else

uninstalling AlphaSubcore on a ship that has only AlphaSubcore would or would not remove the hullmod

AFTER:
installing AlphaSubcore on a ship without modifications would make the hullmod show everywhere.

uninstalling AlphaSubcore from a ship that has only AlphaSubcore removes the hullmod.

The rest of the cases when we had other exotics involved aren't relevant here since that situation was
already addressed properly, and still works fine.