Fix HullmodExoticHandlerTests
The problem was that the runningFromRefitScreen() was directly calling into Starsector API
which returns 'null' for everything while running tests, so all tests were failing with NullPointerException.

Fixed it by decoupling the "are we in refit screen" by delegating to the ne StarsectorAPIInteractor, which
also has the capability to set it to run in "test mode" when it will return a predefined value set by the method.

So, the "Flow" tests need to be added but, not super necesary at the moment, or at all really. Would be great but nah.