Add a warning when trying to install HullmodExotics from Refit screen
Things were extended a bit, and the warning shows up good however, due to making the
checks whether we're running from ExoticaTech screen or Refit screen a bit stricter,
it breaks a certain usecase:

- install NanotechArmor on one child module
- install EqualizerCore on another child module
- install AlphaSubcore from exoticatech screen
- remove AlphaSubcore from exoticatech screen

The AlphaSubcore will not be removed from the child modules.

Need to check whether this actually works fine on the base branch, which probably *never* uses
strict work mode to begin with, since CoreUI tabs are remembered even when they're not showing.

So the (old) typical:
- enter Refit
- press Escape
- enter exotica tech screen

surely still counts as being in "refit".