Get rid of the "remove exoticatech hullmod only if we don't have any more mods" idea
Similar to the 'installThisHullmodExoticToFleetMembersVariant' which had to get rid of it's
"being smart", so does the 'unapplyExoticHullmodAndRemoveExoticaAndHullmod' method too.

But unlike with the install one, the ShipModifications *still* contained the exotica in them
when we reached this method, so we had to remove it from the proper one.

So, with the debugger, the "problematic steps" finished fine and everything ended up being fine.
The AlphaSubcore went away, and OP values reverted properly.

If they don't otherwise, we need synchronization...