Add a warning when trying to install HullmodExotics from Refit screen (#47)
Things were extended a bit, and the warning shows up good however, due to making the checks whether we're running from ExoticaTech screen or Refit screen a bit stricter, it breaks a certain usecase:

- install NanotechArmor on one child module
- install EqualizerCore on another child module
- install AlphaSubcore from exoticatech screen
- remove AlphaSubcore from exoticatech screen

The AlphaSubcore will not be removed from the child modules.

Need to check whether this actually works fine on the base branch, which probably *never* uses strict work mode to begin with, since CoreUI tabs are remembered even when they're not showing.

So the (old) typical:
- enter Refit
- press Escape
- enter exotica tech screen

surely still counts as being in "refit".

UPDATE:
Besides tightening the "runningFromRefit()" and "runningFromExoticaScreen()" methods, this PR also adds synchronization to ShipModLoader.

The "problematic" usecase steps identified above no longer repro, and the warning on the AlphaSubcore shows when using from Refit screen.

---------


* Add a warning when trying to install HullmodExotics from Refit screen

Things were extended a bit, and the warning shows up good however, due to making the
checks whether we're running from ExoticaTech screen or Refit screen a bit stricter,
it breaks a certain usecase:

- install NanotechArmor on one child module
- install EqualizerCore on another child module
- install AlphaSubcore from exoticatech screen
- remove AlphaSubcore from exoticatech screen

The AlphaSubcore will not be removed from the child modules.

Need to check whether this actually works fine on the base branch, which probably *never* uses
strict work mode to begin with, since CoreUI tabs are remembered even when they're not showing.

So the (old) typical:
- enter Refit
- press Escape
- enter exotica tech screen

surely still counts as being in "refit".

* Get rid of the "remove exoticatech hullmod only if we don't have any more mods" idea

Similar to the 'installThisHullmodExoticToFleetMembersVariant' which had to get rid of it's
"being smart", so does the 'unapplyExoticHullmodAndRemoveExoticaAndHullmod' method too.

But unlike with the install one, the ShipModifications *still* contained the exotica in them
when we reached this method, so we had to remove it from the proper one.

So, with the debugger, the "problematic steps" finished fine and everything ended up being fine.
The AlphaSubcore went away, and OP values reverted properly.

If they don't otherwise, we need synchronization...

* Add synchronization to ShipModLoader

Since I think some raceconditions from the StarsectorAPI might be causing out-of-order
sets/gets, it'd be safer if we synchronized access to the static methods.

Ideally, the providers themselves should have synchronization inside of them, but lets try
with just this, since it seems to work fine so far.

* Minor change to ShipModifications to return the result of removeExotic()